Remove FIXME for Simulated event nullability (more info below)
Simulated synthetic events will be missing some properties if they're
not included in eventData, meaning the non-nullable typings on them
are incorrect.

As opposed to making everything nullable to account for that, we'll
treat those simulated events the same way Mockito treats mock objects,
and just require that any properties being accessed get stubbed in using
the eventData argument.